[CHANGE] Made returning presets dynamic
Added a try catch for the ToDict function in case the object hasn't been parsed correctly or doesn;t exist in which case it will return just a null and send the start state to the client.
Removed the StringBuilder as this will be done in a different place now.